refactor: Add modules sidebar component
This commit adds the modules sidebar component to the project. The component is responsible for displaying a list of modules and their respective links in the sidebar. This addition enhances the navigation experience for users and improves the overall usability of the application.

Refactor the code to add the modules sidebar component.