feat: Improve form submission in assist page
The code changes in `+page.blade.php` improve the form submission in the assist page. The `dinamic-form-to-params` class is now applied to the form, and the input placeholder is updated to "DNI, Nombres, apellidos...". This change enhances the user experience by providing a more descriptive placeholder and ensuring consistent styling for the form.

Recent user commits:
- chore: Remove "page" query parameter in form submission

Recent repository commits:
- Update AssistsController to include query parameter in snSchedules method
- feat: Update AssistsController to include query parameter in snSchedules method
- Merge branch 'main' of https://github.com/daustinn/eda
- feat: Add new route for "Asistencias sin horario" in sidebar
- feat: Add new route for "Asistencias sin horario" in sidebar
- feat: Improve filtering functionality in collaborators page
- feat: Update AssistsController to improve lazy loading of images
- feat: Add overflow-auto to assist page for better scrolling
- feat: Add export functionality for assists