refactor: Update AssistsController.php to handle assist_terminals based on estrategy parameter
This commit modifies the AssistsController.php file to handle the assist_terminals parameter based on the estrategy parameter. Previously, the assist_terminals parameter was only considered when the estrategy parameter was not present. With this change, the assist_terminals parameter will be used if the estrategy parameter is set to 'params'.

This modification ensures that the correct terminals are retrieved based on the chosen estrategy, improving the accuracy of the results. It also enhances the flexibility of the code by allowing different strategies for retrieving terminals.

Note: This commit message has been generated based on the provided code changes and recent commits. Please review and modify it as needed.